cleanup

+++ include/vec_formatter.hpp
@@ -14,30 +14,30 @@ struct std::formatter<Vector<T>>
     char spec_buf[32]{};
     bool isColumn = false;      // [:c / C]
     bool isNormalized = false;  // [:n / N]
-    bool isPolar = false;       // [:p / P]
+    bool isPolar = false;       // [:p / P] - add specified output to rad or deg
 
    public:
-    constexpr auto parse(std::format_parse_context& ctx)  // parses spec format and stores state
+    constexpr auto parse(std::format_parse_context& ctx)
     {
-        auto it = ctx.begin();  // points to first character of spec e.g -> '*'
-        auto end = ctx.end();   // points 1 past the ctx buffer.
+        auto it = ctx.begin();
+        auto end = ctx.end();
 
-        if (it == end)  // stop if empty (when spec is {}).
+        if (it == end)
         {
             return it;
         }
 
         auto type_it = it;
-        while (type_it != end && *type_it != '}')  // advance until '}' to find end.
+        while (type_it != end && *type_it != '}')
         {
             ++type_it;
         }
 
-        auto ranged_end = type_it;  // pointing to last '}'
+        auto ranged_end = type_it;
 
         if (ranged_end != it)
         {
-            auto last = ranged_end - 1;  // pointing to last character before '}'
+            auto last = ranged_end - 1;
             if (*last == 'c' || *last == 'C')
             {
                 isColumn = true;
@@ -55,15 +55,13 @@ struct std::formatter<Vector<T>>
             }
         }
 
-        // the range [it, ranged_end] only contains spec parts now.
-        // Store the spec as a string, in a fixed-size char array.
         spec_len = 0;
         for (auto i = it; i != ranged_end; ++i)
         {
             spec_buf[spec_len++] = *i;
         }
 
-        return type_it;  // returning iterator pointing to '}'
+        return type_it;
     }
 
     auto format(const Vector<T>& c, std::format_context& ctx) const  // read the interal state at runtime
@@ -76,14 +74,14 @@ struct std::formatter<Vector<T>>
         }
         else if (isNormalized)
         {
-            double mag = sqrt(c.x() * c.x() + c.y() * c.y());
+            double mag = std::sqrt(c.x() * c.x() + c.y() * c.y());
             result = std::format("({},{})", c.x() / mag, c.y() / mag);
         }
         else if (isPolar)
         {
-            auto mag = sqrt(c.x() * c.x() + c.y() * c.y());
+            auto mag = std::sqrt(c.x() * c.x() + c.y() * c.y());
             // atan2 internally converts int to double, and then convert to degrees
-            auto angle = atan2(c.y(), c.x()) * (180.0 / std::numbers::pi_v<float>);
+            auto angle = std::atan2(c.y(), c.x()) * (180.0 / std::numbers::pi_v<double>);
 
             result = std::format("({},{} deg)", mag, angle);
         }
@@ -91,11 +89,9 @@ struct std::formatter<Vector<T>>
         {
             result = std::format("({},{})", c.x(), c.y());
         }
-        // reconstruct the format string
+
         std::string fmt = std::string{"{:" + std::string(spec_buf, spec_len) + "}"};
 
-        // takes the build string (fmt) wraps the result as the argument to format, and into the
-        // output iterator (ctx)
         return std::vformat_to(ctx.out(), fmt, std::make_format_args(result));
     }
 };
